fixed cake pattern in dailyNotificationHandler and triggerNotificationHandler
