+++ src/main/scala/stockmonitoringbot/notificationhandlers/TriggerNotificationHandlerComponent.scala
@@ -3,6 +3,10 @@ package stockmonitoringbot.notificationhandlers
 /**
   * Created by amir.
   */
+trait TriggerNotificationHandlerComponent {
+  val triggerNotificationHandler: TriggerNotificationHandler
+}
+
 trait TriggerNotificationHandler {
   def start(): Unit
 }
